feat(discounts): add create, edit, and view usage modals for discount codes
- Implemented CreateDiscountModal for creating new discount codes with validation and random code generation.
- Developed EditDiscountModal for editing existing discount codes, including usage statistics and validation.
- Added ViewUsageModal to display usage history of discount codes, including user details and summary statistics.
- Enhanced user experience with loading states and error handling in all modals.
