+++ src/lib/components/admin/Users/Discounts/EditDiscountModal.svelte
@@ -1,20 +1,31 @@
 <script lang="ts">
 	import { createEventDispatcher, getContext } from 'svelte';
 	import { toast } from 'svelte-sonner';
-	import { createDiscountCode, type CreateDiscountForm } from '$lib/apis/discounts';
+	import {
+		updateDiscountCode,
+		type DiscountCode,
+		type UpdateDiscountForm
+	} from '$lib/apis/discounts';
 	import Modal from '$lib/components/common/Modal.svelte';
+
+	export let discount: DiscountCode;
 
 	const dispatch = createEventDispatcher();
 	const i18n: any = getContext('i18n');
 
 	let loading = false;
-	let form: CreateDiscountForm = {
-		code: '',
-		discount_percent: 10,
-		expiry_date: null,
-		usage_limit: null,
-		active: true
+	let form: UpdateDiscountForm = {
+		code: discount.code,
+		discount_percent: discount.discount_percent,
+		expires_at: discount.expires_at,
+		usage_limit: discount.usage_limit,
+		active: discount.active
 	};
+
+	// Separate field for date input (HTML date input expects YYYY-MM-DD format)
+	let expiryDateString: string = discount.expires_at
+		? new Date(discount.expires_at * 1000).toISOString().split('T')[0]
+		: '';
 
 	// Validation
 	let errors: Record<string, string> = {};
@@ -22,7 +33,7 @@
 	function validateForm() {
 		errors = {};
 
-		if (!form.code.trim()) {
+		if (!form.code?.trim()) {
 			errors.code = 'Discount code is required';
 		} else if (form.code.length < 3) {
 			errors.code = 'Discount code must be at least 3 characters';
@@ -30,7 +41,10 @@
 			errors.code = 'Discount code can only contain letters, numbers, hyphens, and underscores';
 		}
 
-		if (form.discount_percent < 1 || form.discount_percent > 100) {
+		if (
+			form.discount_percent !== undefined &&
+			(form.discount_percent < 1 || form.discount_percent > 100)
+		) {
 			errors.discount_percent = 'Discount percentage must be between 1 and 100';
 		}
 
@@ -38,13 +52,13 @@
 			errors.usage_limit = 'Usage limit must be at least 1';
 		}
 
-		if (form.expiry_date) {
-			const expiryDate = new Date(form.expiry_date);
+		if (expiryDateString) {
+			const expiryDate = new Date(expiryDateString);
 			const today = new Date();
 			today.setHours(0, 0, 0, 0);
 
 			if (expiryDate <= today) {
-				errors.expiry_date = 'Expiry date must be in the future';
+				errors.expires_at = 'Expiry date must be in the future';
 			}
 		}
 
@@ -58,101 +72,102 @@
 
 		loading = true;
 		try {
-			// Convert empty strings to null for optional fields
+			// Convert empty strings to null for optional fields and set expires_at from date string
 			const submitData = {
 				...form,
-				code: form.code.trim().toUpperCase(),
-				expiry_date: form.expiry_date || null,
+				code: form.code?.trim().toUpperCase(),
+				expires_at: expiryDateString ? dateStringToTimestamp(expiryDateString) : null,
 				usage_limit: form.usage_limit || null
 			};
 
-			await createDiscountCode(localStorage.token, submitData);
-			toast.success('Discount code created successfully');
-			dispatch('created');
+			await updateDiscountCode(localStorage.token, discount.id!, submitData);
+			toast.success('Discount code updated successfully');
+			dispatch('updated');
 		} catch (error) {
-			console.error('Failed to create discount code:', error);
-			toast.error(typeof error === 'string' ? error : 'Failed to create discount code');
+			console.error('Failed to update discount code:', error);
+			toast.error(typeof error === 'string' ? error : 'Failed to update discount code');
 		} finally {
 			loading = false;
 		}
-	}
-
-	function generateRandomCode() {
-		const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
-		let result = '';
-		for (let i = 0; i < 8; i++) {
-			result += chars.charAt(Math.floor(Math.random() * chars.length));
-		}
-		form.code = result;
-		validateForm();
 	}
 
 	function closeModal() {
 		dispatch('close');
 	}
 
-	// Auto-format expiry date
+	// Auto-format expiry date and convert to timestamp
 	function formatExpiryDate() {
-		if (form.expiry_date) {
-			const date = new Date(form.expiry_date);
-			form.expiry_date = date.toISOString().split('T')[0];
+		if (expiryDateString) {
+			form.expires_at = dateStringToTimestamp(expiryDateString);
+		} else {
+			form.expires_at = null;
 		}
 	}
+
+	// Helper function to convert date string to Unix timestamp
+	const dateStringToTimestamp = (dateString: string | null): number | null => {
+		if (!dateString) return null;
+		return Math.floor(new Date(dateString + 'T23:59:59').getTime() / 1000);
+	};
+
+	// Remove the reactive statement since we handle date conversion directly
 </script>
 
 <Modal size="md" on:close={closeModal}>
-	<div class="flex justify-between dark:text-gray-300 px-5 pt-4 pb-1">
-		<div class="text-lg font-medium self-center flex items-center space-x-2">
-			<svg class="w-5 h-5 text-blue-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-				<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 4v16m8-8H4"
-				></path>
-			</svg>
-			<span>{$i18n.t('Create Discount Code')}</span>
-		</div>
-		<button class="self-center" on:click={closeModal}>
-			<svg
-				xmlns="http://www.w3.org/2000/svg"
-				viewBox="0 0 20 20"
-				fill="currentColor"
-				class="w-5 h-5"
+	<div class="bg-white dark:bg-gray-900 rounded-lg p-6">
+		<div
+			class="flex items-center justify-between border-b border-gray-200 dark:border-gray-700 pb-4 mb-6"
+		>
+			<div class="flex items-center space-x-2">
+				<svg class="w-5 h-5 text-blue-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+					<path
+						stroke-linecap="round"
+						stroke-linejoin="round"
+						stroke-width="2"
+						d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z"
+					></path>
+				</svg>
+				<span class="text-lg font-semibold text-gray-900 dark:text-white"
+					>{$i18n.t('Edit Discount Code')}</span
+				>
+			</div>
+			<button
+				type="button"
+				on:click={closeModal}
+				class="text-gray-400 hover:text-gray-600 dark:hover:text-gray-300"
 			>
-				<path
-					d="M6.28 5.22a.75.75 0 00-1.06 1.06L8.94 10l-3.72 3.72a.75.75 0 101.06 1.06L10 11.06l3.72 3.72a.75.75 0 101.06-1.06L11.06 10l3.72-3.72a.75.75 0 00-1.06-1.06L10 8.94 6.28 5.22z"
-				/>
-			</svg>
-		</button>
-	</div>
-
-	<div class="px-5 pb-4">
+				<svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+					<path
+						stroke-linecap="round"
+						stroke-linejoin="round"
+						stroke-width="2"
+						d="M6 18L18 6M6 6l12 12"
+					></path>
+				</svg>
+			</button>
+		</div>
 		<form on:submit|preventDefault={submitHandler} class="space-y-4">
 			<!-- Discount Code -->
 			<div>
 				<label for="code" class="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1">
 					{$i18n.t('Discount Code')} *
 				</label>
-				<div class="flex space-x-2">
-					<input
-						id="code"
-						type="text"
-						bind:value={form.code}
-						on:input={() => {
+				<input
+					id="code"
+					type="text"
+					bind:value={form.code}
+					on:input={() => {
+						if (form.code) {
 							form.code = form.code.toUpperCase();
-							validateForm();
-						}}
-						placeholder="e.g. SAVE20"
-						class="flex-1 px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 {errors.code
-							? 'border-red-500'
-							: ''}"
-						required
-					/>
-					<button
-						type="button"
-						on:click={generateRandomCode}
-						class="px-3 py-2 text-xs border border-gray-300 dark:border-gray-600 rounded-md bg-gray-50 dark:bg-gray-800 text-gray-700 dark:text-gray-300 hover:bg-gray-100 dark:hover:bg-gray-700"
-					>
-						{$i18n.t('Generate')}
-					</button>
-				</div>
+						}
+						validateForm();
+					}}
+					placeholder="e.g. SAVE20"
+					class="w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 {errors.code
+						? 'border-red-500'
+						: ''}"
+					required
+				/>
 				{#if errors.code}
 					<p class="mt-1 text-sm text-red-600 dark:text-red-400">{errors.code}</p>
 				{/if}
@@ -197,16 +212,16 @@
 				<input
 					id="expiry_date"
 					type="date"
-					bind:value={form.expiry_date}
+					bind:value={expiryDateString}
 					on:change={formatExpiryDate}
 					on:input={validateForm}
 					min={new Date().toISOString().split('T')[0]}
-					class="w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 {errors.expiry_date
+					class="w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-md bg-white dark:bg-gray-700 text-gray-900 dark:text-white focus:ring-2 focus:ring-blue-500 {errors.expires_at
 						? 'border-red-500'
 						: ''}"
 				/>
-				{#if errors.expiry_date}
-					<p class="mt-1 text-sm text-red-600 dark:text-red-400">{errors.expiry_date}</p>
+				{#if errors.expires_at}
+					<p class="mt-1 text-sm text-red-600 dark:text-red-400">{errors.expires_at}</p>
 				{/if}
 				<p class="mt-1 text-xs text-gray-500 dark:text-gray-400">
 					{$i18n.t('Leave empty for no expiry date')}
@@ -238,6 +253,27 @@
 				<p class="mt-1 text-xs text-gray-500 dark:text-gray-400">
 					{$i18n.t('Leave empty for unlimited usage')}
 				</p>
+			</div>
+
+			<!-- Usage Information -->
+			<div class="p-3 bg-gray-50 dark:bg-gray-800 rounded-lg">
+				<h4 class="text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">
+					{$i18n.t('Usage Statistics')}
+				</h4>
+				<div class="grid grid-cols-2 gap-4 text-sm">
+					<div>
+						<p class="text-gray-500 dark:text-gray-400">{$i18n.t('Times Used')}</p>
+						<p class="font-semibold text-gray-900 dark:text-white">{discount.used_count}</p>
+					</div>
+					<div>
+						<p class="text-gray-500 dark:text-gray-400">{$i18n.t('Created At')}</p>
+						<p class="font-semibold text-gray-900 dark:text-white">
+							{discount.created_at
+								? new Date(discount.created_at * 1000).toLocaleDateString()
+								: '-'}
+						</p>
+					</div>
+				</div>
 			</div>
 
 			<!-- Active Status -->
@@ -286,9 +322,9 @@
 								d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
 							></path>
 						</svg>
-						{$i18n.t('Creating...')}
+						{$i18n.t('Updating...')}
 					{:else}
-						{$i18n.t('Create Discount Code')}
+						{$i18n.t('Update Discount Code')}
 					{/if}
 				</button>
 			</div>
